server/services/ml/requirements.txt
google-genai
pillow
python-dotenv
pymongo[srv]
fastapi
uvicorn
python-multipart